d-modem: add resampling from 8kHz to 9.6kHz (fixes V.90)
The softmodem is running at 9600Hz, the SIP stack is running at 8000Hz,
which leads to noise in both directions when communicating with the
softmodem. This makes it possible, to establish V.90 connections
(maximum bitrate I was able to achieve was a stable 37333 for a
connection lasting over 15 minutes).

The resampler can probably be tuned to even more improve this situation
(or replaced with another one).